Added GIFStructs.h/.cpp for GIF-related data types
