+++ src/CMakeLists.txt
@@ -3,6 +3,7 @@ add_library(${ZARKS_LIB_NAME}
     color.cpp
     GaussField.cpp
     GIF.cpp
+    GIFStructs.cpp
     Image.cpp
     Map.cpp
     Mat3.cpp
